:truck: moves chart files to chart folder
