+++ components/pages/file-repository/charts/DataTypesChart/DATA_TYPES_CHART.gql
