+++ components/pages/file-repository/charts/DataTypesChart/index.tsx
@@ -20,8 +20,8 @@
 import SimpleBarChart from '../SimpleBarChart';
 import DATA_TYPES_CHART from './DATA_TYPES_CHART.gql';
 import { useQuery } from '@apollo/react-hooks';
-import { FileRepoFiltersType } from '../utils/types';
-import useFiltersContext from '../hooks/useFiltersContext';
+import { FileRepoFiltersType } from '../../utils/types';
+import useFiltersContext from '../../hooks/useFiltersContext';
 
 type DataTypesChartData = {
   file: {
