+++ components/pages/file-repository/charts/ProgramBarChart/PROGRAMS_CHART.gql
