+++ components/pages/file-repository/charts/SimpleBarChart/index.tsx
