+++ components/pages/file-repository/charts/SimpleBarChart/stories.tsx
