+++ components/pages/file-repository/index.tsx
@@ -29,9 +29,9 @@ import { styled } from 'uikit';
 import { dummyData } from 'components/pages/file-repository/FileTable/dummyData';
 import useAuthContext from 'global/hooks/useAuthContext';
 import FacetPanel from './FacetPanel';
-import FileBarChart from './DataTypesChart';
-import PrimarySiteBarChart from './PrimarySiteBarChart';
-import ProgramBarChart from './ProgramBarChart';
+import FileBarChart from './charts/DataTypesChart';
+import PrimarySiteBarChart from './charts/PrimarySiteBarChart';
+import ProgramBarChart from './charts/ProgramBarChart';
 import StatsCard from './StatsCard';
 
 import { FiltersProvider } from './hooks/useFiltersContext';
